Update dependencies, Improved auth screen UI
