+++ lib/main.dart
@@ -1,32 +1,27 @@
 import 'package:flutter/material.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
-import 'login_page.dart';
+import 'screens/login_screen.dart';
 
-void main()async{
-
-await Firebase.initializeApp(
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized(); // Ensure flutter bindings are initialized
+  await Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
-);
-
- runApp(const MyApp());
- 
+  );
+  runApp(const MyApp());
 }
-
 
 class MyApp extends StatelessWidget {
   const MyApp({super.key});
 
- @override
+  @override
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Flutter Login',
       theme: ThemeData(
         primarySwatch: Colors.blue,
       ),
-      home: LoginPage(),
+      home: LoginScreen(),
     );
   }
 }
-
-
