+++ pubspec.yaml
@@ -40,6 +40,7 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  logger: ^2.0.2+1
 
 dev_dependencies:
   flutter_test:
